requirements.txt
fastapi
uvicorn
python-dotenv
apscheduler
requests
pydantic
python-multipart
httpx
openai-agents